Refactor to have a more flexible test harness
